Extract battle UI facade
